backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.23
alembic==1.12.1
pyodbc==5.0.1
aioodbc==0.4.0
celery==5.3.4
redis==5.0.1
python-socketio==5.9.0
pydantic>=2.7.4
pydantic-settings>=2.1.0
httpx>=0.25.2
gitpython==3.1.40
cryptography==41.0.7
openai==1.3.5
anthropic==0.7.1
langgraph==0.2.60
langchain-core==0.3.28
docker==7.0.0
slowapi==0.1.9
croniter==2.0.1
python-dotenv==1.0.0
aiosqlite==0.19.0
ruff==0.1.8
pytest==7.4.3
pytest-asyncio==0.23.2
